Fix task card
