+++ frontend/src/pages/ProjectDetailPage/ui/CurrentBoard/Column/TaskCard/index.ts
@@ -1,0 +1,1 @@
+export { TaskCard } from "./TaskCard"
